moved cmdline and ply files to conform with the include directory structure. cleaned up the CMakeLists.txt
