+++ graspitConfig.cmake.in
@@ -1,6 +1,6 @@
-# - Config file for the FooBar package
+# - Config file for the GRASPIT package
 # It defines the following variables
-#  GRASPIT_INCLUDE_DIRS - include directories for FooBar
+#  GRASPIT_INCLUDE_DIRS - include directories for GRASPIT
 #  GRASPIT_LIBRARIES    - libraries to link against
 #  GRASPIT_EXECUTABLE   - the bar executable
  
@@ -10,10 +10,10 @@ set(GRASPIT_INCLUDE_DIRS "@GRASPIT_INCLUDES@")
  
 # Our library dependencies (contains definitions for IMPORTED targets)
 if(NOT TARGET graspit_simulator AND NOT GRASPIT_BINARY_DIR)
-  include("${GRASPIT_CMAKE_DIR}/GRASPITTargets.cmake")
+  include("${GRASPIT_CMAKE_DIR}/graspitTargets.cmake")
 endif()
  
-# These are IMPORTED targets created by FooBarTargets.cmake
+# These are IMPORTED targets created by second_endpointTargets.cmake
 set(GRASPIT_LIBRARIES graspit)
 set(GRASPIT_EXECUTABLE graspit_simulator)
